requirements.txt
keras
tensorflow
progressbar2
numpy
matplotlib
sklearn
pandas